# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Syra API",
    "version": "1.0.0",
    "contact": {
      "email": "support@syraa.fun"
    },
    "x-guidance": "Free routes (security: []) include /info and /prediction-game/health. Paid routes return HTTP 402 until settled via x402 (Solana/Base USDC) — pay an offer from the 402 body accepts array and retry with PAYMENT-SIGNATURE or X-PAYMENT. Docs: https://docs.syraa.fun",
...
  },
  "servers": [
    {
      "url": "https://api.syraa.fun"
    }
  ],
...
  "paths": {
...
    "/indicator": {
      "get": {
        "tags": [
          "Technical indicators (x402)"
        ],
        "summary": "OHLCV technical indicators — combine RSI, MACD, EMA, Bollinger, and 20+ more in one agent-readable call",
        "operationId": "getIndicator",
        "security": [
          {
            "x402": []
          }
        ],
        "x-payment-info": {
          "protocols": [
            {
              "x402": {}
            },
            {
              "mpp": {
                "method": "",
                "intent": "",
                "currency": ""
              }
            }
          ],
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.1"
          }
        },
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "BTCUSDT"
            }
          },
          {
            "name": "source",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "binance"
            }
          },
          {
            "name": "interval",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "1h"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 200
            }
          },
          {
            "name": "series",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "indicators",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "rsi"
            },
            "description": "Comma-separated ids e.g. rsi,macd. Per-indicator params: rsi.period=21"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "402": {
            "description": "Payment required — x402 v2 (Payment-Required header + JSON body with accepts). Retry with PAYMENT-SIGNATURE or X-PAYMENT.",
            "headers": {
              "Payment-Required": {
                "description": "x402 v2 encoded payment requirements",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests — burst (25 req / 10s per IP) or sustained (100 req / 60s per IP) limit exceeded. Wait for the duration in `Retry-After` (seconds) before retrying.",
            "headers": {
              "Retry-After": {
                "description": "Seconds the client must wait before retrying.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "message"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Too many requests. Please slow down."
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Technical indicators (x402)"
        ],
        "summary": "Technical indicators (POST, combinable)",
        "description": "Same as GET /indicator with JSON body for complex multi-indicator requests.",
        "operationId": "postIndicator",
        "security": [
          {
            "x402": []
          }
        ],
        "x-payment-info": {
          "protocols": [
            {
              "x402": {}
            },
            {
              "mpp": {
                "method": "",
                "intent": "",
                "currency": ""
              }
            }
          ],
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.1"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "symbol": {
                    "type": "string"
                  },
                  "source": {
                    "type": "string"
                  },
                  "interval": {
                    "type": "string"
                  },
                  "limit": {
                    "type": "integer"
                  },
                  "series": {
                    "type": "boolean"
                  },
                  "indicators": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "402": {
            "description": "Payment required — x402 v2 (Payment-Required header + JSON body with accepts). Retry with PAYMENT-SIGNATURE or X-PAYMENT.",
            "headers": {
              "Payment-Required": {
                "description": "x402 v2 encoded payment requirements",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests — burst (25 req / 10s per IP) or sustained (100 req / 60s per IP) limit exceeded. Wait for the duration in `Retry-After` (seconds) before retrying.",
            "headers": {
              "Retry-After": {
                "description": "Seconds the client must wait before retrying.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "message"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Too many requests. Please slow down."
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  },
...
  "components": {
    "securitySchemes": {
      "x402": {
        "type": "apiKey",
        "in": "header",
        "name": "PAYMENT-SIGNATURE",
        "description": "x402 v2 payment proof (also accepts X-PAYMENT header). Obtain by paying an offer from the HTTP 402 response."
      },
      "SyraApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Legacy internal routes only — openapi-listed free and x402 routes do not use this."
      }
    }
  }
}